tools/mgmt-tester: Fix "Remove Ext Advertising" case
This fixes the expected behavior of the following test case:
"Remove Ext Advertising - Success 1", when the fail status
was erroneously overwritten with success:

<log snip>
  New Advertising Removed event received
  Test condition complete, 2 left
  HCI Command 0x2039 length 6
  Invalid parameter size for HCI command
Remove Ext Advertising - Success 1 - test failed
  HCI Command 0x203c length 1
  HCI Command 0x2039 length 2
  Test condition complete, 1 left
  Remove Advertising (0x003f): Success (0x00)
  Test condition complete, 0 left
Remove Ext Advertising - Success 1 - test passed

The expected HCI command to disable extended advertisement for a
non-zero number of sets should contain the specified number the
advertising sets.
In this particular test case, number of sets is set to 1 and,
as a result, the expected HCI command should be:
"00      | 01             | 01 00 00 00"
 disable | number of sets | set info

Also, to avoid false positives/negatives, skip exmination of HCI
commands after the test conditions are met.