monitor: Work-around memory leak warning
Work-around this warning by making the so pointer global.

Error: RESOURCE_LEAK (CWE-772): [#def29] [important]
monitor/jlink.c:87:3: alloc_fn: Storage is returned from allocation function "dlopen".
monitor/jlink.c:87:3: var_assign: Assigning: "so" = storage returned from "dlopen(jlink_so_name[i], 1)".
monitor/jlink.c:95:2: noescape: Resource "so" is not freed or pointed-to in "dlsym".
monitor/jlink.c:96:2: noescape: Resource "so" is not freed or pointed-to in "dlsym".
monitor/jlink.c:97:2: noescape: Resource "so" is not freed or pointed-to in "dlsym".
monitor/jlink.c:98:2: noescape: Resource "so" is not freed or pointed-to in "dlsym".
monitor/jlink.c:99:2: noescape: Resource "so" is not freed or pointed-to in "dlsym".
monitor/jlink.c:100:2: noescape: Resource "so" is not freed or pointed-to in "dlsym".
monitor/jlink.c:101:2: noescape: Resource "so" is not freed or pointed-to in "dlsym".
monitor/jlink.c:102:2: noescape: Resource "so" is not freed or pointed-to in "dlsym".
monitor/jlink.c:103:2: noescape: Resource "so" is not freed or pointed-to in "dlsym".
monitor/jlink.c:104:2: noescape: Resource "so" is not freed or pointed-to in "dlsym".
monitor/jlink.c:116:2: leaked_storage: Variable "so" going out of scope leaks the storage it points to.
114|
115|	/* don't dlclose(so) here cause symbols from it are in use now */
116|->	return 0;
117|   }
118|