plugin: do not fail when external plugins are disabled
As per the way the old `configure.ac` script was written, it would
always define the controlling variable for external builds. This is not
the expected behaviour as variables not expanded should really be
`#undef`.

Address the issue by checking properly the variable before using it, and
if it is not defined, define locally to be 0 (a.k.a. disabled), to
respect expected behaviour.