configure.ac: check for stdio.h
This fixes a configure failure for readline.h with slibtoolize which
depends on HAVE_STDIO_H being defined. With GNU libtoolize this check is
implicit and with slibtoolize it will fail instead.

  error: unknown type name 'FILE'

Since bluez depends on stdio.h itself there is no reason to not check
for it explicitly.

Gentoo-Issue: https://bugs.gentoo.org/950467