obex: Add supported-features tag in MAP Client Connect Request
This change is required for below PTS testcase -

1. MAP/MCE/MFB/BV-06-C
Verify that the MCE sends its MapSupportedFeatures in the OBEX Connect
request if the MSE declares support for the feature MapSupportedFeatures
in Connect Request in its SDP record.

If Server's SDP record contains the field 'MapSupportedFeatures in Connect
Request' as supported then include the supported features
apparam in obex connect request.

Tested-by: Frédéric Danis <frederic.danis@collabora.com>