main: Fix scanning alert no. 21
Comparison of narrow type with wide type in loop condition
Comparison between of type uint16_t and of wider type size_t.

In a loop condition, comparison of a value of a narrow type with a value
of a wide type may result in unexpected behavior if the wider value is
sufficiently large (or small). This is because the narrower value may
overflow. This can lead to an infinite loop.