transport: Create media request before calling suspend
Creating the media request before suspend can result in
a situation where media_request_reply is not called as
the suspend action for a profile may not depend on the
callback of the stop operation.

This is the case for ASHA profile as devices are not
expected to acknowledge the receipt of a stop command.

Moving media_request_create before media_transport_suspend
should not change the behaviour of release.