obexd: Remove --enable-threads option
Threading support is always enabled in dbus since version 1.8 released
in 2014:
• libdbus always behaves as if dbus_threads_init_default() had been called
  (thread-safety by default)
so remove the obsolete option.

Note that we also require glib 2.36, which is newer than the version of
glib with GThread support built-in:
[Since 2.32]
This function is no longer necessary. The GLib threading system is
automatically initialized at the start of your program.