avdtp: fix not adding stream to global streams list
stream_new() shall be used to create new avdtp_stream, so that they can
be found from global list.

Fix one case where it was not used.  Fixes crash when unregistering lsep
and then disconnecting, due to stream for the removed lsep not being
cleaned up when SEP is freed.

Fixes: https://github.com/bluez/bluez/issues/1325