mesh: main.conf/cfgmod-server: allow to enable proxy support
The proxy settings in the config file and in the configuration server
only affect the _relaying_ between GATT and ADV. Provisioning via
PB-GATT and connecting to the GATT proxy service is always possible
(MshPRT_v1.1, section 4.2.12).

The actual relaying of network messages and beacons is implemented in
the next patches.