device: Add SyncBroadcast method in device1 interface
As a BIS sink role, BlueZ currently defaults to synchronizing
with the first broadcast source device discovered by the le discovery
procedure. This behavior may not align with user expectations,
as it removes control over the target device selection.

This patch introduces a new SyncBroadcast method to the device1 interface,
allowing users to explicitly choose which broadcast source device to
synchronize with.

Signed-off-by: Ye He <ye.he@amlogic.com>