device: allow selecting target broadcast device for synchronization
When multiple broadcast devices are present, bluez currently
defaults to synchronizing with the first one encountered during
the LE discovery procedure. This behavior may not align with
user expectations.

This patch introduces a new SyncBroadcast method in device1,
allowing applications to select the desired target from cached
broadcast devices according to their own criteria — for example,
choosing the device with the strongest RSSI or one matching a
specific set of UUIDs.

Signed-off-by: Ye He <ye.he@amlogic.com>