sixaxis: Fix official DualShock 4 controllers not being handled
The devices changed names, and those do not match the expected name in
our list. Ignore the "Vendor" portion of the device name before matching
it.

This is most likely the result of official DS4 controllers now being
handled by the hid-playstation driver rather than the
community-supported hid-sony.

This fixes DS4 controllers not initiating cable pairing when turned on.