shared/io: add helper for asynchronous shutdown() with SO_LINGER
Add io_glib_shutdown_linger() for socket shutdown with wait for
remote ACK via SO_LINGER. E.g. wait for L2CAP Disconnect Rsp.

We don't want to block the main loop for the linger timeout, so call
shutdown() in a separate thread, as socket API seems to provide only the
blocking way to do it.

Implement it with Glib source API, as that's more convenient for the
AVDTP plugin that needs this.