obexd: Always flush transfer status changes
When transferring multiple files in the same session, obexd reuses the
same obex_transfer object to communicate transfer state to agents.
Because of that every transfer except the first one starts with its
status being "completed". If in addition to that, if the file size is
0 bytes, then the transfer status will change to "active" and then
immediately to "completed", but only change to "completed" will be
signaled through D-Bus. Such status change from "completed" to
"completed" is invalid and can cause issues with some agents. To
resolve the issue, always signal status changes, and set status to
"queued" before starting the transfer.

Fixes: https://github.com/bluez/bluez/issues/1683