shared/gmap: deactivate service or remove feature chrc if role removed
When role is removed, the associated feature chrc must be removed too as
it's then Excluded in spec.  If all roles are removed, deactivate the
service instead to avoid consuming ATT handles.

Reserve exactly the right number of handles for the service.

Also add null checks to the feature set routines for convenience.